+++ frontend/app/features/login/styles/login.module.css
@@ -30,4 +30,10 @@
   .button {
     width: 100%;
   }
+
+  .disabledLink {
+    color: gray;
+    cursor: not-allowed;
+    text-decoration: none;
+  }
