+++ frontend/app/features/password-reset/styles/notFound.module.css
@@ -1,0 +1,13 @@
+.notFoundContainer {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    height: 100vh;
+    text-align: center;
+  }
+  
+  button {
+    margin-top: 1rem;
+  }
+
