+++ frontend/pages/reset-password.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import ResetPasswordForm from "@password-reset/components/ResetPasswordForm";
+import ResetPasswordForm from "@password-reset/components/ResetPasswordRequestForm";
 
 export default function ResetPassword() {
   return <ResetPasswordForm />;
