+++ frontend/pages/update-password.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import ResetPasswordPage from "@features/password-reset/components/Reset-Password";
+import ResetPasswordPage from "@features/password-reset/components/ResetPasswordForm";
 
 export default function ResetPassword() {
   return <ResetPasswordPage />;
